requirements.txt
motor~=3.5.1
python-dotenv~=1.0.1
passlib~=1.7.4
python-jose~=3.5.0
slowapi~=0.1.9
pydantic[email]~=2.10.4
pytest
httpx
pytest-asyncio
qrcode
fastapi~=0.116.1
starlette~=0.47.3
pymongo~=4.8.0
uvicorn~=0.35.0
bcrypt